feat : dailyList 공용컴포넌트로 변경하기 위한 코드수정
