+++ src/containers/DailyPage/components/List.tsx
@@ -1,12 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import ListContents from './ListContents';
-import ListHeader from './ListHeader';
+import ListContents from '../../../components/daily/ListContents';
+import ListHeader from '../../../components/daily/ListHeader';
+import { DailyBoard } from 'src/dummy/dailyInfo';
+import { DailyBoardType } from 'src/types/board.type';
+import { CustomPagination } from 'src/components/CustomPagination';
+
 export function List() {
+  const [activePage, setActivePage] = useState<number>(1);
+  const DailyBoardPiece: DailyBoardType[] = [];
+  const pieceNumber = 10;
+  DailyBoard.map((data, index) => {
+    if (
+      index < (activePage - 1) * pieceNumber ||
+      index > activePage * pieceNumber - 1
+    ) {
+      return;
+    }
+    DailyBoardPiece.push(data);
+  });
   return (
     <Container>
       <ListHeader />
-      <ListContents />
+      <ListContents
+        activePage={activePage}
+        pieceNumber={pieceNumber}
+        dailyData={DailyBoardPiece}
+      />
+      <CustomPagination
+        activePage={activePage}
+        itemsCountPerPage={pieceNumber}
+        totalItemsCount={DailyBoard.length}
+        onChange={e => setActivePage(e)}
+      />
     </Container>
   );
 }
